refactor(test): cast goal state to float

+++ tests/mock_env.py
@@ -15,7 +15,7 @@ class MockEnv(gym.Env):  # pylint: disable=abstract-method
         self.observation_space = Box(high=1, low=-1, shape=(4,), dtype=np.float32)
         action_dim = self.config["action_dim"]
         self.action_space = Box(high=1, low=-1, shape=(action_dim,), dtype=np.float32)
-        self.goal = np.zeros(self.observation_space.shape)
+        self.goal = np.zeros(self.observation_space.shape, dtype=np.float32)
         self.state = None
 
     def reset(self):
@@ -33,4 +33,4 @@ class MockEnv(gym.Env):  # pylint: disable=abstract-method
 
     def reward_fn(self, state, action, next_state):
         # pylint: disable=missing-docstring,unused-argument
-        return torch.norm(next_state - torch.from_numpy(self.goal).float(), dim=-1)
+        return torch.norm(next_state - torch.from_numpy(self.goal), dim=-1)
